requirements.txt
streamlit
pypdf2
langchain
tiktoken
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
